hotfix:replace name area unit and moved filter-by-prop component to filter-by-prop-list component
