+++ src/components/filters/filter-by-prop-list/filter-by-prop/interface.ts
@@ -1,7 +1,7 @@
 import { PropsWithChildren } from 'react';
 import { FieldErrors } from 'react-hook-form';
 
-import FormValues from '../types';
+import FormValues from '../../types';
 
 interface FilterByPropProps extends PropsWithChildren {
   filterName: string;
